Can't have the cake
Sadly, the garbage collection solution falls apart if the expression of
the partially erased duplication node loops. This happens, for example,
in the program below:

    (main) =
      let Y = λf [[λr [f [r r]]] [λr [f [r r]]]]
      let Z = λz λs z
      let f = [Y λslow λn [n [λx x] λpred[[slow pred] [slow pred]]]]
      [f Z]

Here, the y-combinator is used to implement recursion with lambdas, and
will loop with the change. To work properly, ideally, we should call
reduce in another thread, or with some computation limit, or use some
kind of work scheduler that doesn't let it suck all the computing power
of the program. More thought is required, but it is still a very
important aspect to consider.